src/requirements.txt
beautifulsoup4>=4.12.3
dateparser>=1.2.0
flask>=3.0.3
google-api-python-client>=2.146.0
google-auth-httplib2>=0.2.0
google_auth_oauthlib>=1.2.1
gTTS>=2.5.3
matplotlib>=3.9.2
msal>=1.31.0
obd>=0.7.2
openai>=1.50.2
pandas>=2.2.2
pygame>=2.6.0
pyserial>=3.5
PyAudio>=0.2.14
python-dotenv>=1.0.1
pyttsx4>=3.0.15
pytz>=2024.2
requests>=2.32.3
rich>=13.8.1
spacy>=3.7.6
SpeechRecognition>=3.10.4
twilio>=9.3.2